V3.6.1 oh arm64 support (#12994)
* openharmony arm64 support

* update exteral config

+++ native/cocos/platform/interfaces/modules/Device.h
@@ -55,7 +55,7 @@ public:
 
     static uint32_t getInnerWidth();
     static uint32_t getInnerHeight();
-    static uint32_t getWindowHandle();
+    static uintptr_t getWindowHandle();
     /**
      *  Gets the DPI of device
      *  @return The DPI of device.
